- Se agregaron mensajes de error al dejar vacios o no digitar numeros a los textbox (Existencia y Costo)
- Se arreglo el error que provocaba que al guardar o actualizar, en los textbox (Costo y ValorInventario), no se mostraran los decimales.

+++ Migrations/20200618171208_Migracion_Inicial.cs
@@ -13,9 +13,9 @@ namespace PrimerParcial_JoseLuis.Migrations
                     IdArticulo = table.Column<int>(nullable: false)
                         .Annotation("Sqlite:Autoincrement", true),
                     Descripcion = table.Column<string>(nullable: true),
-                    Existencia = table.Column<int>(nullable: false),
-                    Costo = table.Column<int>(nullable: false),
-                    ValorInventario = table.Column<int>(nullable: false)
+                    Existencia = table.Column<double>(nullable: false),
+                    Costo = table.Column<double>(nullable: false),
+                    ValorInventario = table.Column<double>(nullable: false)
                 },
                 constraints: table =>
                 {
